made abstract class and extended fileparser
